Remove duplicated git clone logic.
