+++ utils/git_test.go
@@ -1,4 +1,4 @@
-package builder
+package utils
 
 import (
 	"fmt"
